+++ rme-what4/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for rme-what4
 
+## 0.1.1
+
+* Added support for uninterpreted functions
+
 ## 0.1 -- TBA
 
 * First version.